Added config to script. Added redrowing images on resize.

+++ script.js
@@ -1,13 +1,72 @@
 function MSRY(config) {
+  let _this = this;
+
   this.init = function () {
-    this.rootElement = document.querySelector(config.rootSelector);
-    this.images = document.querySelectorAll(`${config.rootSelector} > img`);
+    this.config = {
+      rootSelector: '#MSRY',
+      settings: {
+        columnsCount: 1,
+      },
+      onScreenWidth: {}
+    };
+
+    // todo: rewrite appending options
+    Object.assign(this.config, config);
+    console.log(this.config);
+
+    this.rootElement = document.querySelector(this.config.rootSelector);
+    if (!this.rootElement) {
+      throw new Error('Wrong selector. Element not found.')
+    }
+
+    this.imagesArr = document.querySelectorAll(`${this.config.rootSelector} > img`);
     this.deleteChildrenFromParent(this.rootElement);
 
-    console.log(this.images)
+    window.addEventListener('resize', onResizeWindowHandler.bind(this));
+    pasteColumnsAndImages.call(this);
+
+    console.log(this.imagesArr)
 
     function onResizeWindowHandler() {
-      console.log(getAppropriateWidth());
+      this.deleteChildrenFromParent(this.rootElement);
+      pasteColumnsAndImages.call(this);
+    }
+
+    function pasteColumnsAndImages() {
+      let columnsCount;
+      let ap = getAppropriateWidth();
+
+      // counting containers
+      if (ap === 'default') {
+        columnsCount = this.config.settings.columnsCount;
+      } else {
+        columnsCount = this.config.onScreenWidth[ap]?.columnsCount || this.config.settings.columnsCount;
+      }
+
+      console.log(ap, columnsCount)
+
+      // creating containers
+      for (let i = 0; i < columnsCount; i++) {
+        let wrapper = document.createElement('div');
+        wrapper.classList.add('msry_wrapper__imgContainer');
+        let space = `- ${columnsCount * 10}px)`;
+        wrapper.style.width = `calc((100% ${columnsCount > 1 ? space : ""} / ${columnsCount})`
+        this.rootElement.appendChild(wrapper);
+      }
+
+      // pasting images into containers
+      // let containerWithMinHeight = null;
+      let pos = 0;
+      Array.from(this.imagesArr).forEach((image, index) => {
+
+        // this.rootElement.children[index % this.imagesArr.length].appendChild(image)
+        this.rootElement.children[pos].appendChild(image);
+
+        pos++;
+        if (pos === this.rootElement.children.length) {
+          pos = 0;
+        }
+      });
     }
 
     function getAppropriateWidth() {
@@ -23,7 +82,7 @@ function MSRY(config) {
       let appropriateWidth;
       for (let i = 0; i < widthArr.length; i++) {
         if (windowWidth > widthArr[0]) {
-          appropriateWidth = widthArr[0];
+          appropriateWidth = 'default';
         } else if (windowWidth <= widthArr[i]) {
           appropriateWidth = widthArr[i];
         }
@@ -31,9 +90,7 @@ function MSRY(config) {
 
       return appropriateWidth;
     }
-
-    window.addEventListener('resize', onResizeWindowHandler.bind(this))
-  }
+  };
 
   this.deleteChildrenFromParent = function (parentElement) {
     parentElement.innerHTML = "";
@@ -44,9 +101,13 @@ function MSRY(config) {
 
 let msry = new MSRY({
   rootSelector: '#MSRY',
-  defaultSettings: {
-    'columnsCount': 3,
-
+  settings: {
+    columnsCount: 4,
   },
-  onScreenWidth: {'1000': {'columnsCount': 3}, 800: {}, 801: {}, 300: {}, 200: {}}
+  onScreenWidth: {
+    1500: {'columnsCount': 4},
+    1160: {'columnsCount': 3},
+    850: {'columnsCount': 2},
+    600: {'columnsCount': 1}
+  }
 });
